requirements.txt
pandas==0.23.3
psycopg2-binary==2.7.4